+++ benchmarks/bench_gitignore_many_to_many.py
@@ -26,6 +26,8 @@ from benchmarks.match_gitignore import (
 GROUP = "GitIgnore.match_files(): 180 lines, 6.5k files"
 
 
+# Hyperscan backend.
+
 @pytest.mark.benchmark(group=GROUP)
 def bench_hs_r1_block_closure(
 	benchmark: BenchmarkFixture,
@@ -137,6 +139,23 @@ def bench_hs_v1(
 	)
 	benchmark(run_match, spec, cpython_files)
 
+
+# Re2 backend.
+
+@pytest.mark.benchmark(group=GROUP)
+def bench_re2_v1(
+	benchmark: BenchmarkFixture,
+	cpython_files: set[str],
+	cpython_gi_lines_all: list[str],
+):
+	spec = GitIgnoreSpec.from_lines(
+		cpython_gi_lines_all,
+		backend='re2',
+	)
+	benchmark(run_match, spec, cpython_files)
+
+
+# Simple backend.
 
 @pytest.mark.benchmark(group=GROUP)
 def bench_sm_filtered(