+++ benchmarks/hyperscan_gitignore_r2.py
@@ -5,10 +5,10 @@ used in benchmarking, but not included in the released library.
 from __future__ import annotations
 
 from collections.abc import (
-	Callable,
 	Sequence)
 from typing import (
 	Any,
+	Callable,  # Replaced by `collections.abc.Callable` in 3.9.2.
 	Optional,  # Replaced by `X | None` in 3.10.
 	Union)  # Replaced by `X | Y` in 3.10.
 from typing_extensions import (