+++ pathspec/_backends/re2/gitignore.py
@@ -6,9 +6,8 @@ contents and structure are likely to change.
 """
 from __future__ import annotations
 
-from collections.abc import (
-	Callable)
 from typing import (
+	Callable,  # Replaced by `collections.abc.Callable` in 3.9.2.
 	Optional,  # Replaced by `X | None` in 3.10.
 	Union)  # Replaced by `X | Y` in 3.10.